% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{class_membership_probabilities}
\alias{class_membership_probabilities}
\title{Get probabilities of class membership for a set of data instances}
\usage{
class_membership_probabilities(tree, df, target_class)
}
\arguments{
\item{tree}{the tree with classification and probability estimation information, built using the function \code{\link{classification_tree}}}

\item{df}{the dataframe containing the data instances}

\item{target_class}{the target variable's value to be used as the class for which membership probability is sought}
}
\value{
A vector of probabilities.
}
\description{
\code{class_membership_probabilities} returns the probabilities of a set of given instances belonging to a specified target class.
}
\seealso{
\code{\link{rank_instances}}
}
