% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{classification_tree}
\alias{classification_tree}
\title{Build a classification tree}
\usage{
classification_tree(df, target, nxt = "seq", stopc = 1000)
}
\arguments{
\item{df}{the data frame containing the dataset}

\item{target}{the target variable, specified either using a character string or the column index in df}

\item{nxt}{the method of growing the tree, which should have one of the following two values: \code{"seq"} (sequentially, level by level) and \code{"ig"} (by highest IG, node by node)}

\item{stopc}{the value at which three building should stop, representing the \strong{number of levels} when \code{nxt="seq"} and the \strong{number of nodes} when \code{nxt="ig"}}
}
\value{
A dataframe containing a row for each tree node, with the following columns:

      \strong{Parent}: row index of parent node

      \strong{Value}:  the value of the parent's splitting attribute that defines this node

      \strong{Attrib}: the splitting attribute for this node

      \strong{Level}:  the level of the node (starts with 1 for the root node, has value 2 for the root nodes' child nodes etc.)

      \strong{IG}:     the Information Gain achieved by the splitting attribute at this node

      \strong{Split}:  a logical value indicating whether the node has been split or not (some nodes may have a splitting attribute assigned but have IG=0, resulting in them not being split)

      \strong{P_<cls>}: proportion of the target class in the subset represented by the node, one for each of the target classes
}
\description{
\code{classification_tree} builds a classification tree using Information Gain as the splitting criterion
}
\seealso{
\code{\link{plot_class_tree}}
}
