% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{entropy}
\alias{entropy}
\title{Calculate the entropy of a vector}
\usage{
entropy(x)
}
\arguments{
\item{x}{the vector}
}
\value{
The entropy value
}
\description{
\code{entropy} uses Claude Shannon's entropy formula
}
\seealso{
\code{\link{information_gain}}
}
