% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{information_gain}
\alias{information_gain}
\title{Calculate the Information Gain for a set split}
\usage{
information_gain(x, y)
}
\arguments{
\item{x}{the target vector i.e. the entire set of labels (target variable values)}

\item{y}{the splitting vector i.e. the vector of values for the splitting attribute}
}
\value{
The calculated Information Gain (IG)
}
\description{
\code{information_gain} calculates the Information Gain resulting from the splitting of a set. This is the difference between the entropy of the entire set and the weighted entropies of the
}
\seealso{
\code{\link{entropy}} \code{\link{best_information_gain}}
}
