% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{plot_class_tree}
\alias{plot_class_tree}
\title{Plots a classification tree}
\usage{
plot_class_tree(tree, attrib_names, fill_page = FALSE,
  probabilities = FALSE)
}
\arguments{
\item{tree}{the tree in the form of a dataframe like that one created by function \code{\link{classification_tree}}}

\item{attrib_names}{the names of the attributes from the dataset (the tree dataframe stores these as indices and the names make for a more readable plot)}

\item{fill_page}{fill the page with the diagram}
}
\description{
\code{plot_class_tree} plots a classification tree from data in the format created by \code{\link{classification_tree}}.
}
\seealso{
\code{\link{classification_tree}}
}
