% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{rank_instances}
\alias{rank_instances}
\title{Ranks instances in a data frame using a classification tree}
\usage{
rank_instances(tree, df, target_class)
}
\arguments{
\item{tree}{the tree with classification and probability estimation information, built using the function \code{\link{classification_tree}}}

\item{df}{the dataframe containing the data instances to be ranked}

\item{target_class}{the target variable's value to be used as the class for ordering (i.e. the higher the probability of the instance being in that class, the higher its rank in the list)}
}
\value{
A vector of instance indices in df, ordered by decreasing probability of membership of class \code{top_class}.
}
\description{
\code{rank_instances} ranks the instances in a data set by probability of them belonging to a specified target class.
}
\seealso{
\code{\link{classification_tree}}
}
